Fix HTTP health check defaults
The default values for HTTP health checks were not set as documented.

Additionally the documentation stated a health check protocol `https`
which does not exist. To enable TLS support for HTTP health checks the
protocol `http` has to be used and the `tls` property has to be set to
true.

Closes #173