Refactor load balancer tests (#192)
The current acceptance tests for the Hetzner Cloud Terraform provider
start to become unwieldy. Especially the creation and formatting of HCL
snippets gets increasingly annoying as more and more resources depending
on other resources are involved.

This commit lays the groundwork for a template driven solution to
creating HCL snippets. Instead of repeating an HCL snippet for each test
step over and over again, only a single Go template per resource/data
source is required. Those templates can be combined as necessary to
create the desired HCL for an acceptance test. This makes creating new
acceptance tests easier and less error prone. Furthermore it reduces the
amount of copy and paste.

Additionally this commit introduces a new directory structure for the
code. Each feature now has its own directory within the internal
directory. The goal is to move code implementing resources and data
sources belonging to that feature to the respective directory. Aside
from making the code structure cleaner this also allows for parallel
execution of acceptance tests for the individual features.

Currently only the tests for the load balancers are moved to their final
destination. Other code will be moved in further commits to keep the
overall amount of changes per commit/pull request manageable. Some code
which will not be needed anymore once the new structure is in place is
marked as deprecated.