Un-deprecate network_id property (#238)
Hetzner Cloud Networks require at least one subnet to exist, before a
resource can be attached to them. In Terraform it is therefore necessary
to create such a subnet using the hcloud_network_subnet resource.
However, the network_id property of hcloud_load_balancer_network as well
as hcloud_server_network references the parent hcloud_network resource.
For Terraform they do not depend on the existence of a subnet. Due to
Terraform executing plan steps it deems independent in parallel, this
creates a race-condition between the creation of the
hcloud_network_subnet and the hcloud_load_balancer_network or
hcloud_server_network resources. The request to create the latter fails
if it reaches the Hetzner Cloud API before the subnet was created.

To alleviate this issue the subnet_id property was introduced to the
hcloud_load_balancer_network and hcloud_server_network resources. The
intention behind this property was to create an explicit dependency
between those resources and the existence of a subnet without requiring
the users to use depends_on. Unfortunately this was done in a haste and
lead to some confusion and issues for users (#234, #235, #236).

This commit un-deprecates the network_id property and retains its
original behavior. The current limitations of the subnet_id property are
documented more clearly.