resource_hcloud_server_network: sleep before creating a hcloud_server_network resource
There seems to be a bug in the Hetzner platform when private networks
are attached immediately after the corresponding server has been
created.

Attaching shortly after, or detaching and reattaching seems to work.

This has been confirmed to be a bug in April at
https://github.com/hetznercloud/terraform-provider-hcloud/issues/161#issuecomment-621125799,
but according to the issue history there, there's still no fix deployed.

Simply sleeping some time before creating the `hcloud_server_network`
resource seems to give the platform enough time to create the machine,
so let's use this as a workaround until the bug in the platform has been
fixed.