Add managed certificates
This commit adds support for Hetzner Cloud managed Certificates. Along
with adding support for managed certificates the existing
`hcloud_certificate` resource is renamed to
`hcloud_uploaded_certificate`. This distinction hopefully makes is easy
to decide which type of certificate is used. An alias for
hcloud_certificate is provided, which continues to function like an
uploaded certificate.