`hcloud_firewall` resource: plugin normalized CIDRs silently
The terraform provider normalized the given CIDRs by default, so when a user entered 10.0.0.1/8 (as a sample) the terraform provider normalized it to 10.0.0.0/8. This lead to a state difference after every apply, so that terraform tries to recreate the firewall rules. After this MR we now validate that the given IP is the start of the CIDR block (e.g. 10.0.0.0/8).

Signed-off-by: Lukas Kämmerling <lukas.kaemmerling@hetzner-cloud.de>