Fix nil pointer dereference for PEM certificates
According to the documentation
[`Decode`](https://pkg.go.dev/encoding/pem#Decode) returns nil if no pem
data is found.

This is possible if the certificate contains additional data which is
not part of the certificate. If this data occurs at the end the code
before this commit panics. This commit ensures that such data is
ignored. Data between two certificates in the chain is ignored as well.
Leading data leads to an error being returned.

Closes #359