Fix wrong sticky session value
We did not add the "sticky_sessions" attribute to the httpMap but
instead set it directly on the resource data.

In addition we had a weird way to identify a service belonging to a load
balancer. First we correctly identifed the affected load balancer service
using our faux service ID. Then we then additionally tried to identify the
same service again. This commit removes this strange behavior and uses
the initially found service in the first place.

Closes #366