Fix non iso8601 timestamp in hcloud_image datasource
The created (& deprecated) dates in the hcloud_image datasource where not iso8601 complaint. Instead they had a strange format: 2020-04-23 17:55:14 +0000 +0000 This fixes the issue now. As those attributes are only in a datasource we should be able to change the format, this shouldn't be an issue as all those values are "computed".

Signed-off-by: Lukas Kämmerling <lukas.kaemmerling@hetzner-cloud.de>