Case-insensitive comparison of IPv6 addresses
Our API converts upper-case IPv6 addresses to lower-case. This in turn
leads to a mismatch between what is stored in the Terraform state and
what our API returns. As a result Terraform tries to update a rule,
which does not need updating.

The solution is to convert any IP addresses to lower-case before
Terraform saves them to the state. This can be easily achieved using the
StateFunc field of schema.Schema. However, when the field is used within
a schema element of schema.TypeSet it creates an additional invalid
value. This to be a known bug in Terraform:
https://github.com/hashicorp/terraform-plugin-sdk/issues/160

For our case we can get by with filtering the invalid entry from the
firewall rules before sending any requests to the API.