Fix date format for certificate states
Previously the created date of a certificate was stored as the Go string
representation of a time.Time value. If the time was in UTC this lead to
issues because the time was stored in an invalid format. This behavior
of the time/Time.String function is known and intended:
https://github.com/golang/go/issues/11712

This commit fixes the created date by storing it in RFC3339 format.
Additionally it adds a migration for states containing the invalid
format to the new format.

We had a similar issue for image time stamps. This was fixed in PR #385.
However, PR #385 does not add a migration for the state. In order to
keep the changes isolated this commit does add the migration
retroactively. This will be done in a separate commit.

Closes #421