Make the image property of hcloud_server optional
Users wishing to import existing infrastructure into Terraform may run
into the error described in #490.

In some cases our backend deleted the reference to the image a server
was created from. This leads to the API returning `null` instead of an
image for the server. This is ok according to the API docs for the
object returned by GET /v1/server/{id} resource. In contrast POST
/v1/server requires the image field to be set to a string value.
Therefore the image property of the Terraform hcloud_server resource was
marked as required since the first commit. However, this makes it
impossible to import existing server resources for which we removed the
reference to the image.

This commit makes the image property optional in Terraform and thus
should enable users to import existing infrastructure. However, as soon
as users change any properties of the server that require the server to
be recreated, they will absolutely need to set the image property.
Failing to do so will result in an error.

Closes #490