Implement hcloud_firewall_attachment resource (#500)
The hcloud_firewall_attachment resource allows to work around the issue
described in #336. Using this resource it is possible to first create a
server and a firewall before attaching the server to the firewall. This
in turn allows the server's public IP addresses to be used in the
firewall rules.

The drawback of this resource is that it requires a workaround to ensure
a server is protected by a firewall right from the beginning. Without
using the workaround there is a small amount of time where the server is
reachable directly from the internet. The workaround is documented
along with the `hcloud_firewall_attachment` resource.

Closes #336