ci: allow unlimited concurrency on integration tests
The `concurrency` keyword automatically cancels any pending jobs for the
same key, so we can not easily use it to build a queue of pending integration
tests and only run one at a time.

For now we decided to allow an unlimited number of concurrent jobs and see
if this causes issues.