fix: error when deleting hcloud_primary_ip with auto_delete
This error sometimes happened if the primary_ip was deleted faster in the API
than in Terraform.

This sometimes caused a flaky e2e test `TestPrimaryIPResource_with_server`,
and was excerbarated by the recent changes made to the server delete code,
where we now wait for the IP to report that is is not assigned, before allowing
any further resource deletions.