fix(server): make sure that each network block is unique
Prior to this fix, multiple network blocks referencing the same network id
on the `hcloud_server` resource were allowed.

This did not cause immediate issues, but only one would be active at a time,
causing a flip-flop of the active configuration with a diff on every apply.

This commit adds some validation to the resource to make sure that every
network block references a unique `network_id`.