fix: load_balancer_target break when lb is recreated
This fixes a bug where the `hcloud_load_balancer_target` resource is
not deleted before a `hcloud_load_balancer` is re-created.

Without this, the loadbalancer will be deleted but the plan does include
the creation of a new hcloud_load_balancer_target, which causes some weird
issues because it suddently disappears from the state or crashes.