fix: firewall_attachment break when firewall is recreated
This fixes a bug where the `hcloud_firewall_attachment` resource is
not deleted before a `hcloud_firewall` is re-created.

Without this, the firewall will be deleted but the plan does include
the creation of a new firewall attachment, which causes some weird
issues because it suddently disappears from the state or crashes.