fix(server): migration to explicit primary IPs fails
When migrating a server from the previously used implicit primary IPs to
using the `hcloud_primary_ip` resources and specifying those in the
`hcloud_server.public_net` argument failed, because the state for
the `public_net` was not correctly populated. This caused the servers to
require an unnecessary re-attach of the primary ips.

Fixes #577