docs: mention that we only accept the location name as attribute (#595)
For some resources it was ambigous if the `location` attribute would
accept the name or id of the location. In practice the terraform provider
only works with the location name, and this was already stated e.g. for
`hcloud_server`.

This commit updates all resource docs to explicitly mention that the
attribute must be the location name.