fix(lb): lb_target breaks when load-balancer is deleted in API
The `hcloud_load_balancer_target` resource breaks when the load-balancer
it is referencing is deleted outside of Terraform:

    hcloud_load_balancer_target.main_target: Refreshing state... [id=lb-srv-tgt-26538015-985324]
    ╷
    │ Error: load balancer 985324: not found

This happened because when trying the get the current state of the resource
it retrieves the load-balancer from the API to find the correct target.
The error handling for this case returned an error back to terraform,
instead of marking the resource as deleted when the load-balancer could not
be found.

This is fixed by switching to a pre-defined error `errLoadBalancerNotFound`
which we can explicitly check for, similar to how it was done with
`errLoadBalancerTargetNotFound` previously.