fix: state is missing resources when rate limit is reached #604 (#621)
If the rate limit is reached we return an error and exit the
reconciliation. This can lead to resources that are actually created, but
were not added to the terraform state based on this process:

1. Resource is created, rate limit was not reached
2. Further API requests are made, usually to wait for async actions
3. Resource is persisted to the state by calling `resource.SetID()`

If any of the calls in (2) fail, the resource will never be added to the
state, but will be created and incur costs in the cloud. This could either
lead to orphaned resources, or to failures in subsequent runs if e.g.
a server with the same name is created again.

We can fix this by calling `resource.SetID()` before making any further API
calls. This way, terraform knows that the resource exists and can retry
any actions on the next run if necessary.