fix(server): avoid recreate when using official image by ID
The hcloud_server resource would always try to use the image name for
the image attribute, even if the user explicitly specified the ID. This
causes the resource to be recreated on every plan, as the desired image
(ID) would be different from the one in the state (name). This only
affected images with a name, so the official os images like
"ubuntu-22.04".

Fixes #657