fix: failed to load plugin schemas
The schemas between the Plugin SDK Provider and the Plugin Framework
Provider have different values. This is not allowed, as it causes
inconstency, and the terraform-provider stops on startup with an error
if this is detected.

The "token" is not truly required in the old provider (or the new),
because we fallback to the HCLOUD_TOKEN environment variable and have
additional validation that verifies it is set.