test(e2e): reduce flakiness in firewall attachment tests (#837)
The firewall endpoints responds with 500 server_error regularly when
being called too many times in parallel. This is annoying as it fails
our integration test suite and requires lengthy retries.

Moving the tests to run in series reduces the amount of parallel calls
to these endpoints. This was previously already done for the firewall
resource and datasource tests.

Sample failure:
https://github.com/hetznercloud/terraform-provider-hcloud/actions/runs/7495395337/job/20405373073?pr=836#step:6:1149