ci: fix coverage of e2e tests not reported
By default "go test" only considers the coverage of code in the same
package as the test. As our e2e tests are located in different packages
to the actual resources, this always resulted in 0% coverage.

I also added codecov flags to e2e & unit tests so we can differentiate
between the coverage in codecov UI.