fix(server): unable to attach two new networks
The current validation fails if a user tries to attach two networks that
are created in the same Apply run to a server. In this case, the
ResourceDiff has both IDs set to `0`, as that value is unknown at plan
time. The current validation fails with error:

    Error: server is only allowed to be attached to each network once: 0

We skip the validation in that case for now. The Terraform SDK v2 does
not give us enough insights to figure out if both refer to the same
resource or if they reference different network resources. We should
revisit this validation after migrating to the Plugin Framework, as it
gives us a lot more details about the proposed plan, and we might have
enough info to give a definitive yay/nay to the user.