fix(ssh-key): data inconsistency with empty label objects
When we implemented the Label helper we added some extra logic that
handled the difference between the `label` attribute default (null) and
the Hetzner Cloud API default (`{}`). This worked well in our test
cases, but breaks if you pass an empty object to the attribute:

    When applying changes to hcloud_ssh_key.this, provider
    "provider[\"registry.terraform.io/hetznercloud/hcloud\"]" produced
    an unexpected new value: .labels: was cty.MapValEmpty(cty.String),
    but now null.

We have now fixed this by setting the default of the labels field to
an empty object to match the return value of the API. With this, we no
longer need the workaround to handle null labels in Terraform config.

Fixes #921