fix(load-balancer): automatic import of target adds conflicting fields
The fields `ip` and `use_private_ip` are marked as conflicting because
`use_private_ip` only takes effect with label selector and server id
targets.

So far we have always set `use_private_ip` to whatever the API returned.
But this causes issues with the command `terraform plan
-generate-config-out=` as it writes all attributes we set in the state
to a TF file. For an LB Target IP this now causes a conflict because
both fields are set.

This is fixed by only setting the field when the target type is not IP.

Closes #961.