feat(server-type): mark included_traffic attribute as deprecated (#963)
The field is deprecated in the API and will be set to `null` in the API
starting on 2024-08-05. This will show up as `0` in the terraform
attribute.

There is currently no alternative for this field in Terraform, as the
pricing information is not being passed through. If you rely on this,
please open an issue to request this with your use case.

Learn more about this change on the [Hetzner Cloud
Changelog](https://docs.hetzner.cloud/changelog#2024-07-25-cloud-api-returns-traffic-information-in-different-format).