fix(primary-ip): conflict when deleting IP (#994)
There sometimes happens a conflict in terraform 1.9+ as it handles the
state update before a delete differently. Both the server resource as
well as the Primary IP resource try to shut down the server, even though
the Primary IP should not as it is no longer assigned to the server.

This causes our e2e tests to fail.

This fixes the regression by explicitly checking that the server still
has the IP assigned before trying to detach it.