fix: remove server network interface workaround (#1021)
Remove a server workaround that was used to fix a private network issue.

The problem that this workaround tried to solve was a timing issue. When
the workaround was implemented, the action returned by the `attach to
network` endpoint was marked as success without waiting for the network
definition to be applied on the host. Today, the actions returned by the
`attach to network` endpoint do wait for the network definition to be
applied on the host.

The workaround can therefor be removed with a certain confidence.

Supersede #1020 

Fixes #1019

Related to #552 and #550

---------

Co-authored-by: sameer mene <sameer.mene@infovista.com>
Co-authored-by: Julian Tölle <julian.toelle@hetzner-cloud.de>