docs: setup auto generated docs (#1046)
Those changes are build on top of #1044. This PR configure the auto
generated docs.

The modification in the files copied within #1044 are the result of the
changes in this PR. Therefore, this PR only explode the docs files into
templates and example files and does not modify any files in /docs.

Follows https://github.com/hashicorp/terraform-plugin-docs to generate
the provider documentation.