fix: abort when data transformation errors
In the previous code any error diagnostics in the data transformation
code for nested objects were not checked before making API calls with
the faulty data. We are not aware of any actual issues. This validation
is just out of caution and protects against bugs we may introduce in
the future.