refactor: add function to wait for all actions to settle (#1268)
Adds a new `hcloudutil.SettleActions()` helper that uses
`ActionClient.WaitForFunc` to wait until all actions have failed or
completed.

Any actions that have failed are then presented as nice error
diagnostics, similar to how we already do it with synchronous API
errors.

If the error returned from `ActionClient.WaitForFunc` matches a
cancelled context, we log the actions that we are still waiting for.

I refactored all framework resources to use this. As its is using the
framework `diag.Diagnostics` type, it is not easily compatible with
SDKv2 resources.