feat(storage-box): add actions to hcloud_storage_box resource (#1251)
Add the actions for `storage_boxes` to the `hcloud_storage_box`
resource.

- https://docs.hetzner.cloud/reference/hetzner#storage-box-actions-change-protection
- https://docs.hetzner.cloud/reference/hetzner#storage-box-actions-change-type
- https://docs.hetzner.cloud/reference/hetzner#storage-box-actions-reset-password
- https://docs.hetzner.cloud/reference/hetzner#storage-box-actions-update-access-settings
- https://docs.hetzner.cloud/reference/hetzner#storage-box-actions-disable-snapshot-plan
- https://docs.hetzner.cloud/reference/hetzner#storage-box-actions-enable-snapshot-plan

Related to #1144.

Notably missing is [Rollback Snapshot](https://docs.hetzner.cloud/reference/hetzner#storage-box-actions-rollback-snapshot).
I was not sure how to best implement this "imperative" one-off action.
There is no way for us to check if it needs to be done, and snapshots
are automatically deleted after a while, making it unsuitable for
terraform. [Terraform `actions`](https://developer.hashicorp.com/terraform/language/v1.14.x/invoke-actions?product_intent=terraform) might be a better fit for this, but they are still in beta in Terraform
and [not supported by OpenTofu](https://github.com/opentofu/opentofu/issues/3309).