feat(storage-box): stop ignoring changes to ssh keys and replace resource instead (#1296)
The current behaviour is not how most other terraform resources work. In
general Terraform providers trust that the user reads the proposed plan
and understands what they mean. We also have no special protection for
other resources that force them to be re-created, like "user_data" or
"ssh_keys" on servers.

To aid users I added an example that shows how to set the attribute and
how to "safeguard" the resource against accidental deletion.

Another reason to remove this behaviour, is that it is not possible to
import resources right now, as Terraform does accept our empty list in
face of the users configured SSH Keys as an acceptable state. See #1287
for this error.

---

This is a **breaking change** for the `hcloud_storage_box`. This is
accepted in a minor release as the resource is marked as experimental
(#1285). I will add a warning to the release notes about this.