fix(zone): using variable for primary nameservers causes error
When evaluating the plan, terraform does multiple rounds of validation,
with more and more of the resource attributes known. At first, the
variable is unknown, so our custom validation shows `Error: Required
attribute`, even though the variable is actually configured.

We can skip checking for unknown, and rely on terraform calling us again
once the variable is actually known to validate if it is null or not.

Fixes #1302