fix(storage-box): retry snapshot+subaccount create when locked (#1307)
The API only allows users to create a single snapshot or subaccount for
any storage box at the same time. While it may be uncommon to create
multiple snapshots at the same time, I can see creating many subaccounts
at the same time being a common thing.

This uses the default exponential backoff and limit of 5 retries each,
so if many resources are "queued", the user may still run into issues.

This was previously already reviewed and merged in #1284, but
accidentally removed in a rebase of the `storage-boxes` branch before it
was merged to `main`. This should reduce the likelihood of #1297.