test: remove old UpgradePluginFramework tests
These tests are added when we migrate a Resource/Datasource to the
terraform plugin framework, and they validate that there is no diff when
updating the provider version.

These tests are no longer necessary after the new version is released,
and can be removed to save on test execution time.