feat: deprecate datacenter in `primary ips` and `servers`
Deprecate datacenter attributes in the following resources/datasources:

- `hcloud_server`
- `data.hcloud_server`
- `data.hcloud_servers`
- `hcloud_primary_ip`
- `data.hcloud_primary_ip`
- `data.hcloud_primary_ips`

Instead of the `datacenter` attribute, please use the location
attribute. This already existed for Servers, and was newly added for
Primary IPs.

The datacenter attribute will be set to an empty string when it is no
longer returned from the API. This is currently planned for after 1 July
2026.

See https://docs.hetzner.cloud/changelog#2025-12-16-phasing-out-datacenters.